IOS-7424: Adapt code to new MEGAContactRequestList.h

+++ MEGAData/Repository/Account/AccountRepository.swift
@@ -78,7 +78,7 @@ final class AccountRepository: NSObject, AccountRepositoryProtocol {
     }
     
     func incomingContactsRequestsCount() -> Int {
-        sdk.incomingContactRequests().size.intValue
+        sdk.incomingContactRequests().size
     }
     
     func relevantUnseenUserAlertsCount() -> UInt {
